requirements.txt
beautifulsoup4==4.9.3
nltk==3.6.3
pandas==1.3.3
Pillow==8.3.2
requests==2.26.0
scikit-learn==0.24.2
tqdm==4.62.3